Rename core package
