+++ pkg/core/required_annotation.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package wave
+package core
 
 import (
 	appsv1 "k8s.io/api/apps/v1"
